chara_fur: DrawFur GetTextureIndex cast fix (+0.122% unit)
Remove redundant unsigned short cast on GetTextureIndex(1) return value.
GetTextureIndex returns 'short' (sign-extending), so comparing directly
with -1 is correct and eliminates the extra clrlwi zero-extend instruction
that was mismatching the target.

+++ src/chara_fur.cpp
@@ -1985,7 +1985,7 @@ void CChara::CModel::DrawFur(Mtx viewMtx, int shadowPass)
 			TextureMan.SetTexture(GX_TEXMAP0, material->GetTexture(0));
 			int hasExtraTexture;
 			int extraTextureFormat;
-			if (static_cast<unsigned short>(material->GetTextureIndex(1)) != -1) {
+			if (material->GetTextureIndex(1) != -1) {
 				CTexture* extraTexture = material->GetTexture(1);
 				hasExtraTexture = 1;
 				TextureMan.SetTexture(GX_TEXMAP2, extraTexture);
